# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Mycorrhizal CRM API
  description: >
    REST API for Mycorrhizal's contact model. As of WP-71 (P2,
    docs/fork-plan/50-integration-and-rebrand.md), the contact endpoints speak
    the neutral RFC 9553-shaped model (contactmodel.Record/Card) rather than
    the legacy flat Contact DTO. GET /contacts (list) returns a slim
    ContactSummary per item; GET/POST/PUT /contacts/{id} exchange the full
    Card/CRM shape (ContactRecordResponse / ContactRecordInput) — a
    symmetric read/write contract.


    T8 (docs/fork-plan/tickets/16-T8-openapi.md) extended this spec to cover
    the whole REST surface: every route registered by
    backend/routes/routes.go under /api/v1 plus the unversioned /health. The
    drift test (backend/openapi_test.go) enumerates the live Gin router and
    fails on any registered route with no spec entry, so this file cannot go
    stale again.

    Coverage notes:

    - **Auth:** every path under /api/v1 except the public auth group
      (register, login, logout, check-password-strength, password-reset,
      auth/oidc/config) requires a session cookie (`auth_token`, set by login
      and OIDC callback) or an `Authorization: Bearer <token>` header. API
      tokens (mycorrhizal_…) are accepted on the general REST path only when
      their scope is `full`; a `carddav`-scoped token is rejected with 403.
      API tokens cannot access the admin paths at all.
...
      `/auth/oidc/config` is always present.
...
      them.
...
servers:
...
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /contacts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    get:
      summary: Get a contact (full neutral Record/Card)
      operationId: getContact
      responses:
        "200":
          description: The contact.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactRecordResponse"
        "404":
          description: Contact not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update a contact
      operationId: updateContact
      description: Same request/response shape as POST /contacts (symmetric read/write contract).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactRecordInput"
      responses:
        "200":
          description: Updated contact.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactRecordResponse"
        "400":
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Contact not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a contact
      operationId: deleteContact
      description: Soft-deletes the contact and cascade-deletes all its associations (notes, activities links, reminders, edges, memberships, life events, field values, sync links).
      responses:
        "200":
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Contact not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /relationship-edges:
    get:
      summary: List relationship edges
      operationId: listRelationshipEdges
      description: >
        Cursor-paginated (T17). Relationship edges are a hard-delete
        collection — the sync.mode is full_resync: a client re-pulls the
        collection rather than using ?since= (edges have no tombstone).
      parameters:
        - name: cursor
          in: query
          description: Opaque resume token from the previous page's next_cursor.
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, default: 25 }
        - name: order
          in: query
          description: Direction for the (updated_at, id) cursor order.
          schema: { type: string, enum: [asc, desc], default: desc }
        - name: contact_id
          in: query
          description: Filter to edges where this Contact.VCardUID is either endpoint.
          schema: { type: string, format: uuid }
        - name: status
          in: query
          description: Filter by status (confirmed or suggested).
          schema: { type: string, enum: [confirmed, suggested] }
        - name: type
          in: query
          description: Filter by relationship type token (e.g. parent_of).
          schema: { type: string }
      responses:
        "401":
          description: Not authenticated (missing or invalid session cookie or full-scope bearer token).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "200":
          description: A page of edges.
          content:
            application/json:
              schema:
                type: object
                properties:
                  relationship_edges:
                    type: array
                    items:
                      $ref: "#/components/schemas/RelationshipEdge"
                  total: { type: integer }
                  next_cursor:
                    type: string
                    description: Resume token for the next page; empty when there are no more rows.
                  limit: { type: integer }
                  sync:
                    $ref: "#/components/schemas/SyncInfo"
    post:
      summary: Create a relationship edge
      operationId: createRelationshipEdge
      description: >
        Each endpoint is either an existing contact (source_id/target_id, a
        Contact.VCardUID) or a not-yet-existing "thin" contact described
        inline (source_thin/target_thin) — exactly one of the pair per
        endpoint. The server always sets Source=user-confirmed,
        Confidence=1.0, Status=confirmed. **Response is wrapped in a
        `relationship_edge` key** (unlike the update/accept endpoints, which
        return the raw object — a real asymmetry that has burned a frontend
        once).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelationshipEdgeInput"
      responses:
        "201":
          description: Edge created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  relationship_edge:
                    $ref: "#/components/schemas/RelationshipEdge"
        "400":
          description: Validation failed, both/neither endpoint given, or source_id == target_id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: An existing endpoint contact was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /relationship-edges/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get a relationship edge
      operationId: getRelationshipEdge
      responses:
        "200":
          description: The edge (raw, not wrapped).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelationshipEdge"
        "404":
          description: Edge not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update a relationship edge (full replace)
      operationId: updateRelationshipEdge
      description: Status/Source/Confidence are server-derived and never changed by an update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RelationshipEdgeInput"
      responses:
        "200":
          description: The updated edge (raw, not wrapped).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelationshipEdge"
        "400":
          description: Validation failed or endpoint resolution failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Edge or endpoint contact not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a relationship edge
      operationId: deleteRelationshipEdge
      description: Hard delete. Doubles as "reject" for suggested edges.
      responses:
        "200":
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Edge not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /relationship-edges/{id}/accept:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      summary: Accept a suggested relationship edge
      operationId: acceptRelationshipEdge
      description: Only valid when the edge's status is "suggested". Returns the raw updated edge.
      responses:
        "200":
          description: Edge accepted (status now confirmed).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelationshipEdge"
        "404":
          description: Edge not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Edge is not a pending suggestion.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # --- Profile pictures ---
...
  /life-events/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      summary: Get a life event
      operationId: getLifeEvent
      responses:
        "200":
          description: The life event (raw).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifeEvent"
        "404":
          description: Life event not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update a life event (full replace)
      operationId: updateLifeEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LifeEventInput"
      responses:
        "200":
          description: The updated life event (raw).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LifeEvent"
        "400":
          description: Validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Life event (or subject contact) not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete a life event (soft delete; its materialized reminder hard-deletes)
      operationId: deleteLifeEvent
      responses:
        "200":
          description: Deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Life event not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # --- Preferences ---
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: auth_token
      description: The httpOnly session cookie set by POST /login and the OIDC callback.
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        A JWT or a `mycorrhizal_…` API token. On the general REST path a
        carddav-scoped token is rejected with 403 (only `full`-scope tokens
        work); API tokens are rejected on every /admin path entirely.
  schemas:
...
    Error:
      type: object
      description: The standard error envelope produced by apperrors.AbortWithError.
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: true
        request_id: { type: string }
        timestamp: { type: string, format: date-time }
    SyncInfo:
      type: object
      description: >
        T17 sync-mode contract carried by every cursor-paginated list
        response. `mode` is how THIS collection syncs; the two arrays are the
        static map of every syncable collection, so a client can plan a full
        sync without hardcoding endpoint names. Incremental collections are
        soft-deleted, support ?since= change feeds, and drop `total`;
        full_resync collections are hard-deleted, are re-pulled wholesale, and
        keep `total`.
      properties:
        mode: { type: string, enum: [incremental, full_resync] }
        incremental:
          type: array
          items: { type: string }
          description: "Collections that sync incrementally via ?since=: contacts, notes, activities, life_events, preferences."
        full_resync:
          type: array
          items: { type: string }
          description: "Collections that require a full re-pull: circles, households, tags, relationship_edges, field_definitions."
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }
      additionalProperties: true
...
    ContactRecordInput:
      type: object
      description: >
        Request body for POST /contacts and PUT /contacts/{id}. Gender rides
        alongside card/crm since it has no neutral-model home (see
        docs/fork-plan/50-integration-and-rebrand.md's RecordFromContact
        notes).
      properties:
        gender:
          type: string
          enum: [male, female, other, prefer_not_to_say]
        card:
          $ref: "#/components/schemas/Card"
        crm:
          $ref: "#/components/schemas/CRMEnvelope"
        passthrough:
          $ref: "#/components/schemas/Passthrough"
    ContactRecordResponse:
      type: object
      description: Full neutral detail-view response for GET/POST/PUT /contacts.
      properties:
        id: { type: integer }
        uid: { type: string }
        etag: { type: string }
        gender: { type: string }
        card:
          $ref: "#/components/schemas/Card"
        crm:
          $ref: "#/components/schemas/CRMEnvelope"
        passthrough:
          $ref: "#/components/schemas/Passthrough"
        photo: { type: string }
        photo_thumbnail: { type: string }
        archived: { type: boolean }
        notes: { type: array, items: { type: object } }
        activities: { type: array, items: { type: object } }
        reminders: { type: array, items: { type: object } }
...
    RelationshipEdge:
      type: object
      description: >
        A directed edge in the relationship graph; Type describes the source's
        role relative to the target. Only status=confirmed edges are fact;
        suggested edges await review.
      properties:
        id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        source_id: { type: string, format: uuid }
        target_id: { type: string, format: uuid }
        type: { type: string }
        directional: { type: boolean }
        metadata:
          type: object
          additionalProperties: true
        source: { type: string, enum: [user-confirmed, household-inferred, imported, ai-suggested] }
        confidence: { type: number }
        status: { type: string, enum: [confirmed, suggested] }
        sensitivity: { type: string, enum: [normal, private, secret] }
    RelationshipEdgeInput:
      type: object
      required: [type]
      description: >
        Exactly one of source_id/source_thin and one of target_id/target_thin
        per endpoint, enforced by the controller. Source/Confidence/Status are
        server-derived and cannot be set by a client.
      properties:
        source_id: { type: string, format: uuid }
        source_thin:
          $ref: "#/components/schemas/ThinContactInput"
        target_id: { type: string, format: uuid }
        target_thin:
          $ref: "#/components/schemas/ThinContactInput"
        type: { type: string }
        sensitivity: { type: string, enum: [normal, private, secret] }
        metadata:
          type: object
          additionalProperties: true
...
    LifeEvent:
      type: object
      description: A permanent fact about an entity's life. remind=false is omitted on the wire (omitempty).
      properties:
        id: { type: string, format: uuid }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        entity_id: { type: string, format: uuid }
        type: { type: string }
        date:
          $ref: "#/components/schemas/PartialDate"
        description: { type: string, maxLength: 2000 }
        source: { type: string, enum: [user, imported, ai-suggested] }
        related_entity_ids:
          type: array
          items: { type: string, format: uuid }
        remind: { type: boolean }
        deleted:
          type: boolean
          description: "Change-feed tombstone marker (T17): true only via ?since= for a soft-deleted life event."
          default: false
    LifeEventInput:
      type: object
      required: [entity_id]
      properties:
        entity_id: { type: string, format: uuid }
        type: { type: string }
        date:
          $ref: "#/components/schemas/PartialDate"
        description: { type: string, maxLength: 2000 }
        source: { type: string, enum: [user, imported, ai-suggested] }
        related_entity_ids:
          type: array
          items: { type: string, format: uuid }
        remind: { type: boolean }

    # --- Preferences ---
...
    CRMEnvelope:
      type: object
      description: Mycorrhizal-specific data outside any contact-exchange standard.
      properties:
        circles: { type: array, items: { type: string } }
        how_we_met: { type: string }
        work_information: { type: string }
        contact_information: { type: string }
...
    PartialDate:
      type: object
      properties:
        year: { type: integer, nullable: true }
        month: { type: integer, nullable: true }
        day: { type: integer, nullable: true }
        calendarScale: { type: string }
...
    Card:
      type: object
      description: >
        Neutral superset of a contact's standardized data (union of RFC 9553
        JSContact + the IANA vCard-elements registry). See
        docs/fork-plan/20-correspondence.md for the full field-by-field
        mapping to/from vCard 3.0/4.0 and JSContact.
      properties:
        uid: { type: string }
        kind: { type: string, enum: [individual, group, org, location, application, device] }
        language: { type: string }
        prodId: { type: string }
        created:
          $ref: "#/components/schemas/Timestamp"
        updated:
          $ref: "#/components/schemas/Timestamp"
        name:
          $ref: "#/components/schemas/Name"
        nicknames:
          type: array
          items: { $ref: "#/components/schemas/Nickname" }
        organizations:
          type: array
          items: { $ref: "#/components/schemas/Organization" }
        titles:
          type: array
          items: { $ref: "#/components/schemas/Title" }
        emails:
          type: array
          items: { $ref: "#/components/schemas/Email" }
        phones:
          type: array
          items: { $ref: "#/components/schemas/Phone" }
        imppAddresses:
          type: array
          items: { $ref: "#/components/schemas/OnlineService" }
        socialProfiles:
          type: array
          items: { $ref: "#/components/schemas/OnlineService" }
        otherOnlineServices:
          type: array
          items: { $ref: "#/components/schemas/OnlineService" }
        addresses:
          type: array
          items: { $ref: "#/components/schemas/Address" }
        anniversaries:
          type: array
          items: { $ref: "#/components/schemas/Anniversary" }
        speakToAs:
          $ref: "#/components/schemas/SpeakToAs"
        personalInfo:
          type: array
          items: { $ref: "#/components/schemas/PersonalInfo" }
        notes:
          type: array
          items: { $ref: "#/components/schemas/CardNote" }
        keywords:
          type: array
          items: { type: string }
        media:
          type: array
          items: { $ref: "#/components/schemas/Resource" }
        calendars:
          type: array
          items: { $ref: "#/components/schemas/Resource" }
        freeBusyUrls:
          type: array
          items: { $ref: "#/components/schemas/Resource" }
        schedulingAddresses:
          type: array
          items: { $ref: "#/components/schemas/Resource" }
        cryptoKeys:
          type: array
          items: { $ref: "#/components/schemas/Resource" }
        directories:
          type: array
          items: { $ref: "#/components/schemas/Resource" }
        links:
          type: array
          items: { $ref: "#/components/schemas/Resource" }
        contactUris:
          type: array
          items: { $ref: "#/components/schemas/Resource" }
        preferredLanguages:
          type: array
          items: { $ref: "#/components/schemas/LanguagePref" }
        relatedTo:
          type: array
          items: { $ref: "#/components/schemas/Relation" }
        members:
          type: array
          items: { type: string }
        localizations:
          type: object
          additionalProperties: true